feat: transfer function from commands to utils
